Rename handler class
